Added Chandra Village theme, new instrument samples, new voicegroup, and improved sound system
-Added Song
I had to tweak .gitignore so that I could push the .s file for the song into the project, since that's the file in which I do most of my work.

-Extra Instruments
The instruments I've added are some common instruments that aren't in pokeemerald.

-New Voicegroup
This voicegroup is one large voicegroup that has basically every instrument in it, with each instrument mapped to the same location as they'd be mapped in a midi program. This makes it a lot easier to transfer a midi directly into the ROM.

-Sound System Upgrade
Songs can now have up to 12 DirectSound (non 8-bit) notes playing at once, as well as have up to 16 tracks.